Spectral analysis of extremes

Performs a simple spectral density estimation on the indicator series of exceedances above a threshold.

Usage

spectral_analysis_extremes(x, threshold)

Arguments

x

Numeric vector.

threshold

Numeric threshold for exceedances.

Value

List with frequency and spectral density components as returned by [stats::spectrum()].

Examples

x <- rnorm(1000)
spectral_analysis_extremes(x, threshold = 2)
#> $freq
#>   [1] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.010 0.011 0.012
#>  [13] 0.013 0.014 0.015 0.016 0.017 0.018 0.019 0.020 0.021 0.022 0.023 0.024
#>  [25] 0.025 0.026 0.027 0.028 0.029 0.030 0.031 0.032 0.033 0.034 0.035 0.036
#>  [37] 0.037 0.038 0.039 0.040 0.041 0.042 0.043 0.044 0.045 0.046 0.047 0.048
#>  [49] 0.049 0.050 0.051 0.052 0.053 0.054 0.055 0.056 0.057 0.058 0.059 0.060
#>  [61] 0.061 0.062 0.063 0.064 0.065 0.066 0.067 0.068 0.069 0.070 0.071 0.072
#>  [73] 0.073 0.074 0.075 0.076 0.077 0.078 0.079 0.080 0.081 0.082 0.083 0.084
#>  [85] 0.085 0.086 0.087 0.088 0.089 0.090 0.091 0.092 0.093 0.094 0.095 0.096
#>  [97] 0.097 0.098 0.099 0.100 0.101 0.102 0.103 0.104 0.105 0.106 0.107 0.108
#> [109] 0.109 0.110 0.111 0.112 0.113 0.114 0.115 0.116 0.117 0.118 0.119 0.120
#> [121] 0.121 0.122 0.123 0.124 0.125 0.126 0.127 0.128 0.129 0.130 0.131 0.132
#> [133] 0.133 0.134 0.135 0.136 0.137 0.138 0.139 0.140 0.141 0.142 0.143 0.144
#> [145] 0.145 0.146 0.147 0.148 0.149 0.150 0.151 0.152 0.153 0.154 0.155 0.156
#> [157] 0.157 0.158 0.159 0.160 0.161 0.162 0.163 0.164 0.165 0.166 0.167 0.168
#> [169] 0.169 0.170 0.171 0.172 0.173 0.174 0.175 0.176 0.177 0.178 0.179 0.180
#> [181] 0.181 0.182 0.183 0.184 0.185 0.186 0.187 0.188 0.189 0.190 0.191 0.192
#> [193] 0.193 0.194 0.195 0.196 0.197 0.198 0.199 0.200 0.201 0.202 0.203 0.204
#> [205] 0.205 0.206 0.207 0.208 0.209 0.210 0.211 0.212 0.213 0.214 0.215 0.216
#> [217] 0.217 0.218 0.219 0.220 0.221 0.222 0.223 0.224 0.225 0.226 0.227 0.228
#> [229] 0.229 0.230 0.231 0.232 0.233 0.234 0.235 0.236 0.237 0.238 0.239 0.240
#> [241] 0.241 0.242 0.243 0.244 0.245 0.246 0.247 0.248 0.249 0.250 0.251 0.252
#> [253] 0.253 0.254 0.255 0.256 0.257 0.258 0.259 0.260 0.261 0.262 0.263 0.264
#> [265] 0.265 0.266 0.267 0.268 0.269 0.270 0.271 0.272 0.273 0.274 0.275 0.276
#> [277] 0.277 0.278 0.279 0.280 0.281 0.282 0.283 0.284 0.285 0.286 0.287 0.288
#> [289] 0.289 0.290 0.291 0.292 0.293 0.294 0.295 0.296 0.297 0.298 0.299 0.300
#> [301] 0.301 0.302 0.303 0.304 0.305 0.306 0.307 0.308 0.309 0.310 0.311 0.312
#> [313] 0.313 0.314 0.315 0.316 0.317 0.318 0.319 0.320 0.321 0.322 0.323 0.324
#> [325] 0.325 0.326 0.327 0.328 0.329 0.330 0.331 0.332 0.333 0.334 0.335 0.336
#> [337] 0.337 0.338 0.339 0.340 0.341 0.342 0.343 0.344 0.345 0.346 0.347 0.348
#> [349] 0.349 0.350 0.351 0.352 0.353 0.354 0.355 0.356 0.357 0.358 0.359 0.360
#> [361] 0.361 0.362 0.363 0.364 0.365 0.366 0.367 0.368 0.369 0.370 0.371 0.372
#> [373] 0.373 0.374 0.375 0.376 0.377 0.378 0.379 0.380 0.381 0.382 0.383 0.384
#> [385] 0.385 0.386 0.387 0.388 0.389 0.390 0.391 0.392 0.393 0.394 0.395 0.396
#> [397] 0.397 0.398 0.399 0.400 0.401 0.402 0.403 0.404 0.405 0.406 0.407 0.408
#> [409] 0.409 0.410 0.411 0.412 0.413 0.414 0.415 0.416 0.417 0.418 0.419 0.420
#> [421] 0.421 0.422 0.423 0.424 0.425 0.426 0.427 0.428 0.429 0.430 0.431 0.432
#> [433] 0.433 0.434 0.435 0.436 0.437 0.438 0.439 0.440 0.441 0.442 0.443 0.444
#> [445] 0.445 0.446 0.447 0.448 0.449 0.450 0.451 0.452 0.453 0.454 0.455 0.456
#> [457] 0.457 0.458 0.459 0.460 0.461 0.462 0.463 0.464 0.465 0.466 0.467 0.468
#> [469] 0.469 0.470 0.471 0.472 0.473 0.474 0.475 0.476 0.477 0.478 0.479 0.480
#> [481] 0.481 0.482 0.483 0.484 0.485 0.486 0.487 0.488 0.489 0.490 0.491 0.492
#> [493] 0.493 0.494 0.495 0.496 0.497 0.498 0.499 0.500
#> 
#> $spec
#>   [1] 7.964487e-03 3.207036e-02 4.624245e-02 1.047408e-02 1.934916e-02
#>   [6] 8.055226e-04 2.495030e-03 1.260712e-02 3.093146e-03 5.615256e-02
#>  [11] 1.649048e-02 3.501958e-02 1.124205e-01 2.315838e-02 3.513526e-03
#>  [16] 2.382753e-02 5.670470e-03 2.319671e-02 1.601833e-02 1.547568e-02
#>  [21] 2.460234e-02 6.298277e-02 1.173336e-01 9.364895e-03 1.109848e-02
#>  [26] 1.598051e-02 2.989399e-03 1.473110e-02 1.368153e-02 2.163741e-02
#>  [31] 5.407467e-03 2.407501e-02 2.424243e-02 4.256704e-02 3.115826e-02
#>  [36] 6.354290e-02 5.821803e-03 2.986273e-02 1.277115e-02 1.301447e-02
#>  [41] 1.280332e-02 5.251048e-03 1.015479e-03 3.249535e-02 2.916487e-04
#>  [46] 1.636536e-02 4.681475e-02 6.301041e-04 1.926427e-02 6.820136e-03
#>  [51] 1.979311e-03 4.981810e-02 8.356535e-03 8.418998e-04 1.002100e-02
#>  [56] 1.080737e-02 8.857007e-03 1.798275e-03 5.845436e-02 1.519795e-02
#>  [61] 5.169255e-02 1.224630e-02 8.159608e-04 3.074786e-02 1.425065e-02
#>  [66] 1.277470e-02 2.487067e-03 1.417485e-03 1.507087e-03 4.135931e-02
#>  [71] 1.053853e-02 1.792216e-02 6.369731e-02 1.169278e-02 3.887776e-02
#>  [76] 4.756974e-03 1.581881e-02 2.005994e-02 3.884806e-03 1.287094e-04
#>  [81] 1.115611e-02 1.790248e-02 3.201428e-02 3.807669e-03 5.110383e-03
#>  [86] 1.447013e-02 2.516363e-03 1.551251e-03 2.875488e-02 1.323306e-02
#>  [91] 2.792585e-02 4.029164e-02 3.222722e-02 9.871254e-03 4.661852e-02
#>  [96] 4.694084e-02 4.008272e-02 5.354210e-03 3.773885e-03 6.877959e-03
#> [101] 1.065653e-02 8.014093e-03 5.886638e-02 1.109649e-02 6.912044e-02
#> [106] 1.638338e-02 2.896590e-03 4.717887e-02 2.509748e-02 1.414199e-02
#> [111] 1.610569e-02 2.692051e-02 2.746449e-02 8.337236e-04 1.575531e-02
#> [116] 4.312246e-03 2.281085e-02 6.254997e-02 4.019732e-02 1.649751e-02
#> [121] 7.695422e-03 4.282577e-02 8.146457e-03 3.729961e-03 9.481077e-02
#> [126] 3.785995e-02 3.991033e-03 1.880257e-02 1.492038e-03 2.310386e-03
#> [131] 3.340726e-02 2.556323e-02 1.359706e-03 1.373377e-03 3.230739e-02
#> [136] 2.253297e-04 2.869243e-02 2.490946e-02 2.359983e-02 5.838965e-03
#> [141] 8.610461e-02 9.960223e-03 1.187213e-03 4.223897e-02 2.271200e-02
#> [146] 2.596679e-03 6.370933e-02 4.103123e-02 7.085510e-03 4.120530e-02
#> [151] 1.600006e-02 1.089116e-02 1.762086e-03 4.886055e-02 3.624875e-03
#> [156] 4.562748e-03 1.043121e-02 4.071919e-03 1.462688e-02 4.607470e-02
#> [161] 2.574276e-03 1.463516e-02 1.138773e-03 6.464872e-02 1.093467e-02
#> [166] 5.039285e-04 3.540965e-02 3.633353e-03 2.187858e-02 1.396036e-02
#> [171] 3.028046e-02 3.228655e-02 7.388357e-03 1.277899e-02 1.455995e-02
#> [176] 9.933917e-04 1.275460e-02 1.415048e-02 1.445502e-02 1.221372e-02
#> [181] 1.528950e-02 2.351558e-02 2.811300e-03 1.253865e-02 1.982668e-02
#> [186] 8.269946e-03 1.004850e-02 2.980081e-02 1.909592e-02 1.449041e-02
#> [191] 1.022110e-04 1.374987e-02 2.754668e-02 6.326258e-02 4.479103e-03
#> [196] 4.084665e-04 1.585160e-02 3.186105e-02 3.013525e-03 1.995466e-02
#> [201] 2.275523e-02 5.196696e-02 2.012948e-02 3.616063e-03 1.883470e-02
#> [206] 1.811685e-02 4.809314e-03 6.919970e-02 4.300736e-04 6.993657e-03
#> [211] 6.881779e-02 2.752858e-02 2.651779e-02 4.546306e-02 3.438497e-02
#> [216] 4.662703e-02 1.594288e-03 3.856482e-03 1.717746e-03 2.417935e-02
#> [221] 4.820663e-02 6.959914e-03 1.498208e-02 8.105905e-02 9.413403e-04
#> [226] 4.337317e-02 4.381519e-02 1.136309e-03 7.878121e-03 1.565366e-02
#> [231] 2.133299e-02 1.664797e-02 7.455341e-03 2.872599e-02 2.277781e-03
#> [236] 7.686580e-03 5.904672e-02 2.261246e-04 2.616161e-04 2.011560e-02
#> [241] 8.509278e-04 1.801801e-02 3.633183e-02 6.934999e-02 1.177014e-02
#> [246] 1.261423e-02 3.090227e-02 6.851309e-04 1.599490e-02 5.839348e-06
#> [251] 9.157713e-03 1.697702e-03 6.245413e-05 2.367748e-02 5.054609e-02
#> [256] 1.685531e-02 3.536301e-02 2.231015e-02 2.106900e-02 4.075441e-02
#> [261] 1.889090e-03 1.323854e-02 6.428636e-03 4.887435e-02 1.003961e-02
#> [266] 2.445799e-02 8.021387e-02 1.555718e-02 6.443778e-03 1.145496e-02
#> [271] 7.719481e-03 3.426888e-02 6.462590e-03 7.628167e-03 1.624991e-02
#> [276] 1.361064e-02 2.019654e-02 8.749488e-03 4.076882e-02 1.010977e-02
#> [281] 5.457053e-03 4.645517e-03 2.345352e-02 1.222797e-02 7.009626e-02
#> [286] 4.252408e-02 2.094519e-02 5.459268e-03 4.173269e-02 2.279832e-02
#> [291] 4.521192e-02 1.478906e-02 1.614888e-03 4.617242e-03 1.010331e-03
#> [296] 2.736301e-02 6.880825e-02 3.135509e-02 3.901799e-02 2.194210e-03
#> [301] 1.716210e-03 2.974427e-02 5.383101e-04 9.510390e-03 3.047704e-02
#> [306] 4.743889e-03 3.209506e-02 1.885807e-02 1.636874e-03 4.315249e-03
#> [311] 1.983984e-02 1.416176e-02 2.748736e-03 1.903193e-02 2.206338e-02
#> [316] 2.946490e-02 3.429446e-03 2.640777e-02 1.146286e-01 5.739820e-02
#> [321] 3.122071e-02 3.852624e-03 6.558793e-03 3.660741e-03 1.203936e-02
#> [326] 2.742804e-03 1.380611e-03 5.401186e-04 4.445974e-02 8.471721e-03
#> [331] 2.504933e-03 4.533024e-02 3.549241e-02 2.433450e-02 1.832519e-02
#> [336] 4.820589e-03 1.866357e-02 2.211083e-02 1.355079e-02 2.385113e-03
#> [341] 9.787414e-03 1.198088e-02 8.151423e-03 7.536003e-03 7.288030e-03
#> [346] 2.232935e-03 6.984886e-03 1.185225e-03 7.980105e-02 7.714306e-05
#> [351] 4.868675e-03 9.544375e-02 1.442767e-03 3.850104e-03 1.843894e-02
#> [356] 3.417700e-02 3.719242e-02 3.142999e-02 1.411311e-02 1.124297e-02
#> [361] 4.958260e-02 5.771848e-02 2.635748e-02 5.098525e-03 3.269201e-02
#> [366] 1.315741e-02 1.281307e-02 8.192706e-03 1.339539e-03 1.706229e-02
#> [371] 4.568735e-03 4.153681e-02 7.953038e-03 1.834521e-02 3.012078e-02
#> [376] 2.783681e-02 2.577568e-02 1.549092e-02 9.773033e-03 5.445666e-02
#> [381] 1.244512e-02 1.698422e-03 2.182086e-02 6.402295e-02 1.729515e-02
#> [386] 5.486279e-03 4.215700e-03 3.525038e-02 1.218106e-02 4.605560e-02
#> [391] 4.241169e-03 1.391759e-02 2.948879e-02 1.006873e-02 2.856134e-02
#> [396] 4.989475e-02 7.197817e-03 1.139139e-02 2.696439e-02 6.160216e-05
#> [401] 4.342148e-02 9.271040e-03 1.225272e-02 1.068744e-02 4.327120e-03
#> [406] 2.926657e-03 1.392160e-03 5.868457e-02 3.370936e-03 1.423843e-02
#> [411] 7.551834e-02 7.710848e-03 3.254038e-02 4.583993e-02 9.125982e-03
#> [416] 5.159272e-02 1.339761e-02 2.259818e-02 5.561174e-03 2.125969e-03
#> [421] 8.545425e-03 7.088910e-02 8.103603e-03 9.848953e-02 1.542317e-03
#> [426] 3.322805e-02 2.166257e-02 5.814123e-03 2.593529e-02 1.402251e-02
#> [431] 2.539059e-02 1.365734e-02 1.723966e-04 8.633767e-03 2.114777e-02
#> [436] 3.211736e-02 6.671092e-02 9.577628e-03 3.696167e-02 8.489048e-03
#> [441] 5.795348e-03 1.574260e-02 9.492062e-04 2.200365e-02 5.244632e-02
#> [446] 3.127978e-03 6.963089e-03 1.731769e-04 1.061780e-02 8.356383e-03
#> [451] 2.653578e-03 3.263034e-03 5.326538e-03 2.495436e-02 1.508369e-02
#> [456] 6.698329e-03 2.097226e-02 3.759273e-02 6.384157e-03 4.472546e-02
#> [461] 5.004193e-03 1.935441e-03 5.165623e-03 1.982122e-03 5.858297e-04
#> [466] 3.265089e-02 2.444608e-02 5.175175e-03 3.905694e-02 2.829334e-04
#> [471] 6.839572e-03 5.027638e-03 9.827642e-03 2.017949e-02 9.102426e-03
#> [476] 8.454125e-03 2.425956e-02 6.145655e-03 5.228704e-02 1.431594e-02
#> [481] 1.863597e-02 1.050601e-03 1.914841e-02 8.268846e-03 4.934350e-03
#> [486] 5.304197e-03 1.875388e-02 2.344381e-02 1.180967e-02 4.012890e-03
#> [491] 2.528363e-02 1.254029e-02 1.469432e-03 2.790802e-03 4.489427e-03
#> [496] 1.395815e-03 2.240831e-02 2.806563e-02 4.226705e-02 4.838556e-03
#> 
#> $coh
#> NULL
#> 
#> $phase
#> NULL
#> 
#> $kernel
#> NULL
#> 
#> $df
#> [1] 1.79159
#> 
#> $bandwidth
#> [1] 0.0002886751
#> 
#> $n.used
#> [1] 1000
#> 
#> $orig.n
#> [1] 1000
#> 
#> $series
#> [1] "x"
#> 
#> $snames
#> NULL
#> 
#> $method
#> [1] "Raw Periodogram"
#> 
#> $taper
#> [1] 0.1
#> 
#> $pad
#> [1] 0
#> 
#> $detrend
#> [1] TRUE
#> 
#> $demean
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "spec"